+++ lib/router.dart
@@ -19,7 +19,7 @@ import 'package:myecl/settings/router.dart';
 import 'package:myecl/raffle/router.dart';
 import 'package:myecl/tools/middlewares/authenticated_middleware.dart';
 import 'package:myecl/tools/middlewares/deferred_middleware.dart';
-import 'package:myecl/tricount/ui/router.dart';
+import 'package:myecl/tricount/router.dart';
 import 'package:myecl/vote/router.dart';
 import 'package:qlevar_router/qlevar_router.dart';
