+++ lib/settings/providers/module_list_provider.dart
@@ -12,7 +12,7 @@ import 'package:myecl/event/router.dart';
 import 'package:myecl/home/router.dart';
 import 'package:myecl/loan/router.dart';
 import 'package:myecl/raffle/router.dart';
-import 'package:myecl/tricount/ui/router.dart';
+import 'package:myecl/tricount/router.dart';
 import 'package:myecl/vote/router.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -67,10 +67,13 @@ class ModulesNotifier extends StateNotifier<List<Module>> {
     List<String> modulesName = prefs.getStringList(dbModule) ?? [];
     List<String> allSavedModulesName = prefs.getStringList(dbAllModules) ?? [];
     final allModulesName = allModules.map((e) => e.root.toString()).toList();
+    print(allModulesName.map((e) => e,).toList());
     if (modulesName.isEmpty) {
       modulesName = allModulesName;
       saveModules();
     }
+    print(allSavedModulesName.isEmpty);
+    print(!eq.equals(allSavedModulesName, allModulesName));
     if (allSavedModulesName.isEmpty ||
         !eq.equals(allSavedModulesName, allModulesName)) {
       allSavedModulesName = allModulesName;
@@ -97,6 +100,7 @@ class ModulesNotifier extends StateNotifier<List<Module>> {
         }
       }
     }
+    print(modules.map((e) => e.name,).toList());
     for (Module module in toDelete) {
       allModules.remove(module);
     }
