+++ lib/tricount/ui/pages/tricount.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:myecl/tools/ui/widgets/top_bar.dart';
 import 'package:myecl/tricount/tools/constants.dart';
-import 'package:myecl/tricount/ui/router.dart';
+import 'package:myecl/tricount/router.dart';
 
 class TricountTemplate extends HookConsumerWidget {
   final Widget child;
